seperate wrapper implementation
